refactor: rename plugin module to obsidian module
