+++ src/obsidian/obsidian-hack/command.ts
