+++ src/obsidian/obsidian-hack/index.ts
