+++ src/obsidian/obsidian-hack/obsidian-debug-mobile.ts
