+++ src/obsidian/obsidian-hack/plugin.ts
