+++ src/obsidian/vault.ts
